[alpaka] Move the function static objects to .cc files
Split devices.h into host.{h,cc} and alpaka/devices.{h,cc} .
