+++ src/alpaka/AlpakaCore/getDeviceCachingAllocator.h
@@ -7,7 +7,7 @@
 
 #include "AlpakaCore/AllocatorConfig.h"
 #include "AlpakaCore/CachingAllocator.h"
-#include "AlpakaCore/devices.h"
+#include "AlpakaCore/alpaka/devices.h"
 #include "AlpakaCore/getDeviceIndex.h"
 
 namespace cms::alpakatools {
