+++ src/alpaka/test/alpaka/HistoContainer_t.cc
@@ -6,7 +6,8 @@
 #include <random>
 
 #include "AlpakaCore/HistoContainer.h"
-#include "AlpakaCore/devices.h"
+#include "AlpakaCore/alpaka/devices.h"
+#include "AlpakaCore/host.h"
 #include "AlpakaCore/memory.h"
 #include "AlpakaCore/initialise.h"
 #include "AlpakaCore/workdivision.h"
