+++ src/alpaka/test/alpaka/OneHistoContainer_t.cc
@@ -5,7 +5,8 @@
 #include <random>
 
 #include "AlpakaCore/HistoContainer.h"
-#include "AlpakaCore/devices.h"
+#include "AlpakaCore/alpaka/devices.h"
+#include "AlpakaCore/host.h"
 #include "AlpakaCore/initialise.h"
 #include "AlpakaCore/memory.h"
