+++ src/alpaka/test/alpaka/clustering_t.cc
@@ -9,7 +9,7 @@
 #include <set>
 #include <vector>
 
-#include "AlpakaCore/devices.h"
+#include "AlpakaCore/alpaka/devices.h"
 #include "AlpakaCore/initialise.h"
 #include "AlpakaCore/memory.h"
 #include "AlpakaCore/workdivision.h"
